Updated the update checking code to first download the release version when a new release version has been released when the user is the public beta branch. Before it would automatically skip to the next beta branch version. This was incorrect behavior.
The newest release branch version should always be the version that's downloaded when a new branch version is released.

If this situation occurs, the program sets a value in the Registry to trigger an update check the next time the program runs to check for any new public beta branch versions.